Refactoring: Relocating types to more suited locations
