+++ docs/api/aerial/client/__init__.md
@@ -4,135 +4,6 @@
 
 Client functionality.
 
-<a id="cosmpy.aerial.client.__init__.Account"></a>
-
-## Account Objects
-
-```python
-@dataclass
-class Account()
-```
-
-Account.
-
-<a id="cosmpy.aerial.client.__init__.StakingPosition"></a>
-
-## StakingPosition Objects
-
-```python
-@dataclass
-class StakingPosition()
-```
-
-Staking positions.
-
-<a id="cosmpy.aerial.client.__init__.UnbondingPositions"></a>
-
-## UnbondingPositions Objects
-
-```python
-@dataclass
-class UnbondingPositions()
-```
-
-Unbonding positions.
-
-<a id="cosmpy.aerial.client.__init__.Validator"></a>
-
-## Validator Objects
-
-```python
-@dataclass
-class Validator()
-```
-
-Validator.
-
-<a id="cosmpy.aerial.client.__init__.Coin"></a>
-
-## Coin Objects
-
-```python
-@dataclass
-class Coin()
-```
-
-Coins.
-
-<a id="cosmpy.aerial.client.__init__.StakingSummary"></a>
-
-## StakingSummary Objects
-
-```python
-@dataclass
-class StakingSummary()
-```
-
-Get the staking summary.
-
-<a id="cosmpy.aerial.client.__init__.StakingSummary.total_staked"></a>
-
-#### total`_`staked
-
-```python
-@property
-def total_staked() -> int
-```
-
-Get the total staked amount.
-
-<a id="cosmpy.aerial.client.__init__.StakingSummary.total_rewards"></a>
-
-#### total`_`rewards
-
-```python
-@property
-def total_rewards() -> int
-```
-
-Get the total rewards.
-
-<a id="cosmpy.aerial.client.__init__.StakingSummary.total_unbonding"></a>
-
-#### total`_`unbonding
-
-```python
-@property
-def total_unbonding() -> int
-```
-
-total unbonding.
-
-<a id="cosmpy.aerial.client.__init__.Block"></a>
-
-## Block Objects
-
-```python
-@dataclass
-class Block()
-```
-
-Block.
-
-<a id="cosmpy.aerial.client.__init__.Block.from_proto"></a>
-
-#### from`_`proto
-
-```python
-@staticmethod
-def from_proto(block: Any) -> "Block"
-```
-
-Parse the block.
-
-**Arguments**:
-
-- `block`: block as Any
-
-**Returns**:
-
-parsed block as Block
-
 <a id="cosmpy.aerial.client.__init__.LedgerClient"></a>
 
 ## LedgerClient Objects
